Reorganize Parquet code
