+++ pyproject.toml
@@ -35,7 +35,7 @@ implicit_reexport = true
 
 # TODO: fix!
 [[tool.mypy.overrides]]
-module = "fairseq2.data.parquet_tools.*,tests.integration.parquet.*,recipes.parquet.*"
+module = "fairseq2.data.parquet.tools,tests.integration.parquet.*"
 ignore_errors = true
 
 [tool.pytest.ini_options]
