+++ src/fairseq2/data/parquet/dataloader.py
@@ -13,7 +13,7 @@ import pyarrow as pa
 import pyarrow.parquet as pq
 
 from fairseq2.data.data_pipeline import DataPipeline, DataPipelineBuilder
-from fairseq2.data.parquet_tools import (
+from fairseq2.data.parquet.tools import (
     BatchOutputType,
     _TableWrapper,
     _to_real_object,
